main/Faikin: Web controls work
Implement /control URL, which accepts parameters under the same name as /status
produces. This allows to support web panel with minimal efforts, close to
original websocket implementation. We could use BRP-compatible URLs, but this
approach is better because it provides all the settings on the same URL

Some missing BRP control functionality is implemented, but not tested yet

Signed-off-by: Pavel Fedin <pavel_fedin@mail.ru>