main/Faikin: Improve loopback detection
With old version it was possible to trigger a false positive by sending 'MM'
command to the conditioner using mqtt "send" message. The conditioner replies
with 'MFFFF' (yes, non-standard form). It's also possible to make both command
bytes matching by sending 'MF', which causes the same response.

Check for full replica of a sent packet to be on the safe side.

Signed-off-by: Pavel Fedin <pavel_fedin@mail.ru>